webscanner-backend/requirements.txt
absl-py==1.3.0
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.1
Flask==2.3.2
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
intel-openmp==2021.4.0
itsdangerous==2.1.2
jax==0.4.29
Jinja2==3.1.2
joblib==1.4.2
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.9.0
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
mysql-connector-python==8.0.33
networkx==3.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.2
pillow==10.3.0
plotly==5.15.0
protobuf==3.20.3
psycopg2-binary==2.9.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.10.1
six==1.16.0
soupsieve==2.5
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.5.0
tldextract==5.1.2
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
tzdata==2024.1
ucimlrepo==0.0.7
urllib3==1.26.16
Werkzeug==2.3.7
whois==1.20240129.2
wrapt==1.14.1
PyJWT==2.8.0